Fixing the python unit tests
