+++ run_tests.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 echo $DB
-rm ~/.caravel/caravel_unittests.db
+rm ~/.caravel/unittests.db
 rm ~/.caravel/celerydb.sqlite
 rm ~/.caravel/celery_results.sqlite
 rm -f .coverage
