+++ tests/core_tests.py
@@ -7,103 +7,23 @@ from __future__ import unicode_literals
 from datetime import datetime
 import csv
 import doctest
-import json
 import imp
-import os
+import json
 import io
 import unittest
 
-from mock import Mock, patch
 
 from flask import escape
 from flask_appbuilder.security.sqla import models as ab_models
 
 import caravel
 from caravel import app, db, models, utils, appbuilder, sm
-from caravel.models import DruidCluster, DruidDatasource
+from caravel.models import DruidDatasource
 
-os.environ['CARAVEL_CONFIG'] = 'tests.caravel_test_config'
-
-# Disable celery.
-app.config['CELERY_CONFIG'] = None
-app.config['CSRF_ENABLED'] = False
-app.config['PUBLIC_ROLE_LIKE_GAMMA'] = True
-app.config['SECRET_KEY'] = 'thisismyscretkey'
-app.config['SQL_SELECT_AS_CTA'] = False
-app.config['TESTING'] = True
-app.config['WTF_CSRF_ENABLED'] = False
+from .base_tests import CaravelTestCase
 
 BASE_DIR = app.config.get("BASE_DIR")
 cli = imp.load_source('cli', BASE_DIR + "/bin/caravel")
-
-
-class CaravelTestCase(unittest.TestCase):
-
-    def __init__(self, *args, **kwargs):
-        super(CaravelTestCase, self).__init__(*args, **kwargs)
-        self.client = app.test_client()
-
-        utils.init(caravel)
-
-        admin = appbuilder.sm.find_user('admin')
-        if not admin:
-            appbuilder.sm.add_user(
-                'admin', 'admin', ' user', 'admin@fab.org',
-                appbuilder.sm.find_role('Admin'),
-                password='general')
-
-        gamma = appbuilder.sm.find_user('gamma')
-        if not gamma:
-            appbuilder.sm.add_user(
-                'gamma', 'gamma', 'user', 'gamma@fab.org',
-                appbuilder.sm.find_role('Gamma'),
-                password='general')
-
-        alpha = appbuilder.sm.find_user('alpha')
-        if not alpha:
-            appbuilder.sm.add_user(
-                'alpha', 'alpha', 'user', 'alpha@fab.org',
-                appbuilder.sm.find_role('Alpha'),
-                password='general')
-
-        utils.init(caravel)
-
-    def login(self, username='admin', password='general'):
-        resp = self.client.post(
-            '/login/',
-            data=dict(username=username, password=password),
-            follow_redirects=True)
-        assert 'Welcome' in resp.data.decode('utf-8')
-
-    def get_query_by_sql(self, sql):
-        session = db.create_scoped_session()
-        query = session.query(models.Query).filter_by(sql=sql).first()
-        session.close()
-        return query
-
-    def logout(self):
-        self.client.get('/logout/', follow_redirects=True)
-
-    def test_welcome(self):
-        self.login()
-        resp = self.client.get('/caravel/welcome')
-        assert 'Welcome' in resp.data.decode('utf-8')
-
-    def setup_public_access_for_dashboard(self, table_name):
-        public_role = appbuilder.sm.find_role('Public')
-        perms = db.session.query(ab_models.PermissionView).all()
-        for perm in perms:
-            if (perm.permission.name == 'datasource_access' and
-                    perm.view_menu and table_name in perm.view_menu.name):
-                appbuilder.sm.add_permission_role(public_role, perm)
-
-    def revoke_public_access(self, table_name):
-        public_role = appbuilder.sm.find_role('Public')
-        perms = db.session.query(ab_models.PermissionView).all()
-        for perm in perms:
-            if (perm.permission.name == 'datasource_access' and
-                    perm.view_menu and table_name in perm.view_menu.name):
-                appbuilder.sm.del_permission_role(public_role, perm)
 
 
 class CoreTests(CaravelTestCase):
@@ -336,12 +256,6 @@ class CoreTests(CaravelTestCase):
         self.logout()
         return json.loads(resp.data.decode('utf-8'))
 
-    # TODO(bkyryliuk): fix the gamma access, they shouldn't be allowed running queries.
-    # def test_sql_json_no_access(self):
-    #     result = self.run_sql('SELECT * FROM ab_user', 'gamma')
-    #     self.assertEqual('SQL Lab requires the `all_datasource_access` or specific DB permission',
-    #                      result['error'])
-
     def test_sql_json(self):
         data = self.run_sql('SELECT * FROM ab_user', 'admin')
         assert len(data['data']) > 0
@@ -382,14 +296,6 @@ class CoreTests(CaravelTestCase):
         sql = "SELECT first_name, last_name FROM ab_user " \
               "where first_name='admin'"
         self.run_sql(sql, 'admin')
-
-        # TODO(bkyryliuk): fix the permissions
-        # No access if the user is not logged in.
-        # query1_id = self.get_query_by_sql(sql).id
-        # no_access_reps = self.client.get('/caravel/csv/{}'.format(query1_id))
-        # self.assertTrue(
-        #     'SQL Lab requires the `all_datasource_access` or specific DB permission'
-        #     in no_access_reps.data.decode('utf-8'))
 
         query1_id = self.get_query_by_sql(sql).id
         self.login('admin')
@@ -498,115 +404,6 @@ class CoreTests(CaravelTestCase):
         db.session.commit()
         self.test_save_dash('alpha')
 
-SEGMENT_METADATA = [{
-  "id": "some_id",
-  "intervals": ["2013-05-13T00:00:00.000Z/2013-05-14T00:00:00.000Z"],
-  "columns": {
-    "__time": {
-        "type": "LONG", "hasMultipleValues": False,
-        "size": 407240380, "cardinality": None, "errorMessage": None},
-    "dim1": {
-        "type": "STRING", "hasMultipleValues": False,
-        "size": 100000, "cardinality": 1944, "errorMessage": None},
-    "dim2": {
-        "type": "STRING", "hasMultipleValues": True,
-        "size": 100000, "cardinality": 1504, "errorMessage": None},
-    "metric1": {
-        "type": "FLOAT", "hasMultipleValues": False,
-        "size": 100000, "cardinality": None, "errorMessage": None}
-  },
-  "aggregators": {
-    "metric1": {
-        "type": "longSum",
-        "name": "metric1",
-        "fieldName": "metric1"}
-  },
-  "size": 300000,
-  "numRows": 5000000
-}]
-
-GB_RESULT_SET = [
-  {
-    "version": "v1",
-    "timestamp": "2012-01-01T00:00:00.000Z",
-    "event": {
-      "name": 'Canada',
-      "sum__num": 12345678,
-    }
-  },
-  {
-    "version": "v1",
-    "timestamp": "2012-01-01T00:00:00.000Z",
-    "event": {
-      "name": 'USA',
-      "sum__num": 12345678 / 2,
-    }
-  },
-]
-
-
-class DruidTests(CaravelTestCase):
-
-    """Testing interactions with Druid"""
-
-    def __init__(self, *args, **kwargs):
-        super(DruidTests, self).__init__(*args, **kwargs)
-
-    @patch('caravel.models.PyDruid')
-    def test_client(self, PyDruid):
-        self.login(username='admin')
-        instance = PyDruid.return_value
-        instance.time_boundary.return_value = [
-            {'result': {'maxTime': '2016-01-01'}}]
-        instance.segment_metadata.return_value = SEGMENT_METADATA
-
-        cluster = (
-            db.session
-            .query(DruidCluster)
-            .filter_by(cluster_name='test_cluster')
-            .first()
-        )
-        if cluster:
-            db.session.delete(cluster)
-        db.session.commit()
-
-        cluster = DruidCluster(
-            cluster_name='test_cluster',
-            coordinator_host='localhost',
-            coordinator_port=7979,
-            broker_host='localhost',
-            broker_port=7980,
-            metadata_last_refreshed=datetime.now())
-
-        db.session.add(cluster)
-        cluster.get_datasources = Mock(return_value=['test_datasource'])
-        cluster.get_druid_version = Mock(return_value='0.9.1')
-        cluster.refresh_datasources()
-        datasource_id = cluster.datasources[0].id
-        db.session.commit()
-
-        resp = self.client.get('/caravel/explore/druid/{}/'.format(
-            datasource_id))
-        assert "[test_cluster].[test_datasource]" in resp.data.decode('utf-8')
-
-        nres = [
-            list(v['event'].items()) + [('timestamp', v['timestamp'])]
-            for v in GB_RESULT_SET]
-        nres = [dict(v) for v in nres]
-        import pandas as pd
-        df = pd.DataFrame(nres)
-        instance.export_pandas.return_value = df
-        instance.query_dict = {}
-        instance.query_builder.last_query.query_dict = {}
-        resp = self.client.get(
-            '/caravel/explore/druid/{}/?viz_type=table&granularity=one+day&'
-            'druid_time_origin=&since=7+days+ago&until=now&row_limit=5000&'
-            'include_search=false&metrics=count&groupby=name&flt_col_0=dim1&'
-            'flt_op_0=in&flt_eq_0=&slice_id=&slice_name=&collapsed_fieldsets=&'
-            'action=&datasource_name=test_datasource&datasource_id={}&'
-            'datasource_type=druid&previous_viz_type=table&json=true&'
-            'force=true'.format(datasource_id, datasource_id))
-        assert "Canada" in resp.data.decode('utf-8')
 
 if __name__ == '__main__':
     unittest.main()
